requirements.txt
numpy==1.26.4
scikit-image==0.24.0
torch=2.3.1
torchvision==0.18.1
pillow==10.4.0
opencv-python==4.10.0.84
gradio==4.39.0
wandb==0.17.5
tqdm==4.66.4